pyproject.toml
[project]
name = "cse291"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.10.*"
dependencies = [
    "accelerate>=1.11.0",
    "diffusers>=0.35.2",
    "einops>=0.8.1",
    "ftfy>=6.3.1",
    "huggingface-hub[cli]>=0.36.0",
    "hydra-core>=1.3.2",
    "imageio>=2.37.0",
    "imageio-ffmpeg>=0.6.0",
    "kornia==0.8.1",
    "kornia-rs==0.1.9",
    "matplotlib>=3.10.7",
    "numpy>=1.26.4",
    "nvitop>=1.5.3",
    "omegaconf>=2.3.0",
    "open-clip-torch==3.2.0",
    "open3d>=0.19.0",
    "opencv-python>=4.11.0.86",
    "pdbpp>=0.11.7",
    "pillow>=12.0.0",
    "plotly>=5.0.0",
    "pycocotools>=2.0.10",
    "sam-2",
    "setuptools==80.9.0",
    "supervision>=0.26.1",
    "timm==1.0.22",
    "tokenizers>=0.20.3",
    "torch>=2.9.0",
    "torchaudio>=2.9.0",
    "torchvision>=0.24.0",
    "traceanything==0.0.1",
    "viser>=1.0.15",
    "xfuser>=0.4.4",
]

[tool.uv.sources]
sam-2 = { git = "https://github.com/facebookresearch/segment-anything-2.git" }
traceanything = { path = "./TraceAnything" }
